restructure specs a little
