+++ deployables/extension/src/panel/pages/$activeAccountId/transactions/Transactions.spec.ts
@@ -357,6 +357,8 @@ describe('Transactions', () => {
           url: `http://localhost/account/${account.id}`,
         })
       })
+
+      it.todo('respects the currently selected route')
     })
   })
